move out the resolve function
