+++ convert/resolve_test.go
@@ -1,0 +1,3 @@
+package bedgovcf
+
+// TODO write tests
